Schemas names refactor
