+++ src/server/models/user.js
@@ -2,8 +2,8 @@ const dbConnection = require('../dbConnectionProvider').getDbConnection()
 const _ = require('lodash')
 const mongoose = require('mongoose')
 const uniqueValidator = require('mongoose-unique-validator')
-const pendingNotificationsSchema = require('./pendingNotifications')
-const friendsSchema = require('./friends')
+const pendingNotificationsSchema = require('./pendingNotificationsSchema')
+const friendsSchema = require('./friendsSchema')
 const Schema = mongoose.Schema
 
 const DUPLICATE_KEY_ERROR_DESC = 'There was a duplicate key'
